Moved the aerosols function back to creator
