+++ epochal/OIFS/eocene_functions.py
@@ -170,123 +170,6 @@ def albedo(field: xr.Dataset, var=None, lsm_present=None, landsea=None, **kwargs
     print("Eocene land-sea mask applied successfully.")
     print("Combined modification complete, GRIB structure preserved.")
     return field
-    
-
-def aerosols(self):
-    """
-    Convert Herold Eocene aerosol data (kg/kg) to column-integrated mass (kg/m²)
-    on the IFS grid, using the US Standard Atmosphere for density interpolation.
-    """
-
-    print("→ Loading Herold aerosol data")
-    paleoaerfile = os.path.join(self.herold, "herold_etal_eocene_CAM4_BAM_aerosols.nc")
-    aer_paleo = xr.open_dataset(paleoaerfile)
-
-    # Load IFS reference aerosol climatology
-    aerfile_ifs_pd = os.path.join(self.idir, 'oifs', 'ifsdata/aerosol_cams_climatology_43R3a.nc')
-    aer_ifs_pd = xr.open_dataset(aerfile_ifs_pd)
-    aer_ifs_paleo = aer_ifs_pd.copy()
-
-    # Re-bin the Herold aerosol bins
-    aer_paleo_newbin = aer_paleo.copy()
-    aer_paleo_newbin['DST01'] = aer_paleo['DST01']*0.57
-    aer_paleo_newbin['DST02'] = aer_paleo['DST01']*0.39
-    aer_paleo_newbin['DST03'] = aer_paleo['DST02'] + aer_paleo['DST03'] + aer_paleo['DST04']
-    aer_paleo_newbin['SSLT01'] = aer_paleo['SSLT01']*0.57
-    aer_paleo_newbin['SSLT02'] = aer_paleo['SSLT01']*0.39 + aer_paleo['SSLT02'] + aer_paleo['SSLT03']
-    aer_paleo_newbin['SSLT03'] = aer_paleo['SSLT04']
-
-    # Check or generate US Standard Atmosphere
-    ua_file = os.path.join(self.herold, "us_standard_atmosphere_newlevs.nc")
-    if not os.path.exists(ua_file):
-        print("→ Generating US Standard Atmosphere data")
-        # US Standard Atmosphere data (alt, temp, g, pressure, density, mu)
-        data = """
-        -1000   21.50   9.810   11.39   1.347   1.821
-        0       15.00   9.807   10.13   1.225   1.789
-        1000    8.50    9.804   8.988   1.112   1.758
-        2000    2.00    9.801   7.950   1.007   1.726
-        3000    -4.49   9.797   7.012   0.9093  1.694
-        4000    -10.98  9.794   6.166   0.8194  1.661
-        5000    -17.47  9.791   5.405   0.7364  1.628
-        6000    -23.96  9.788   4.722   0.6601  1.595
-        7000    -30.45  9.785   4.111   0.5900  1.561
-        8000    -36.94  9.782   3.565   0.5258  1.527
-        9000    -43.42  9.779   3.080   0.4671  1.493
-        10000   -49.90  9.776   2.650   0.4135  1.458
-        15000   -56.50  9.761   1.211   0.1948  1.422
-        20000   -56.50  9.745   0.5529  0.08891 1.422
-        25000   -51.60  9.730   0.2549  0.04008 1.448
-        30000   -46.64  9.715   0.1197  0.01841 1.475
-        40000   -22.80  9.684   0.0287  0.003996 1.601
-        50000   -2.5    9.654   0.007978 0.001027 1.704
-        60000   -26.13  9.624   0.002196 0.0003097 1.584
-        70000   -53.57  9.594   0.00052  0.00008283 1.438
-        80000   -74.51  9.564   0.00011  0.00001846 1.321
-        """
-        lines = data.strip().split('\n')
-        parsed_data = [list(map(float, line.split())) for line in lines]
-        arr = np.array(parsed_data)
-        altitude = arr[:,0]
-        temperature = 273.15 + arr[:,1]
-        g = arr[:,2]
-        pressure = 1e2*arr[:,3]
-        density = arr[:,4]
-        mu = arr[:,5]
-
-        ds = xr.Dataset(
-            {
-                'temperature': (['altitude'], temperature),
-                'gravity': (['altitude'], g),
-                'pressure': (['altitude'], pressure),
-                'density': (['altitude'], density),
-                'viscosity': (['altitude'], mu)
-            },
-            coords={'altitude': (['altitude'], altitude)}
-        )
-
-        # log-pressure coordinate & interpolate to Herold levels
-        log_pressure = np.log(ds.pressure)
-        ds = ds.assign_coords(log_pressure=log_pressure)
-        ds_new = ds.swap_dims({'altitude':'log_pressure'})
-        new_log_pressure = np.log(aer_paleo.lev)
-        ds_interp = ds_new.interp(log_pressure=new_log_pressure, method='linear')
-        ds_interp.to_netcdf(ua_file)
-        print(f"→ US Standard Atmosphere saved at {ua_file}")
-    else:
-        ds_interp = xr.open_dataset(ua_file)
-
-    # Map Herold -> IFS variable names
-    var_dict = {
-        'Sulfates': "SO4",
-        'Black_Carbon_hydrophilic': "CB1",
-        'Black_Carbon_hydrophobic': "CB2",
-        'Mineral_Dust_bin1': "DST01",
-        'Mineral_Dust_bin2': "DST02",
-        'Mineral_Dust_bin3': "DST03",
-        'Organic_Matter_hydrophilic': "OC1",
-        'Organic_Matter_hydrophobic': "OC2",
-        'Sea_Salt_bin1': "SSLT01",
-        'Sea_Salt_bin2': "SSLT02",
-        'Sea_Salt_bin3': "SSLT03"
-    }
-
-    # Convert to column mass and regrid
-    for varname2 in var_dict:
-        varname1 = var_dict[varname2]
-        print(f"→ Processing {varname1} -> {varname2}")
-        new_var = aer_paleo_newbin[varname1] * ds_interp.density
-        new_var_rg = regrid_dataset(new_var, regrid_to_reference=aer_ifs_paleo[varname2])
-        new_var_rg_int = -new_var_rg.integrate(coord='altitude')
-        aer_ifs_paleo[varname2].data = new_var_rg_int.data.astype('float32')
-
-    # Save Eocene aerosol climatology
-    output_path = os.path.join(self.odir_init, "aerosol_cams_climatology_43R3a_EOCENO.nc")
-    if os.path.exists(output_path):
-        os.remove(output_path)
-    aer_ifs_paleo.to_netcdf(output_path)
-    print(f"→ Eocene aerosol data saved at {output_path}")
-    return output_path
 
 def compute_slope (grib_field, sd_eoc, a=4.376786e-05, b=2.476405e-04):
     """
